fix(ci): line-limit — split hosted validation part_005 over 1500 lines
